requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.10
fa2_modified==0.3.10
Flask==3.0.3
Flask-SocketIO==5.3.6
graphistry==0.33.8
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
networkx==3.3
numpy==2.0.0
packaging==24.1
palettable==3.3.3
pandas==2.2.2
pyarrow==16.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
scipy==1.14.0
simple-websocket==1.0.0
six==1.16.0
squarify==0.4.3
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
websocket-server==0.6.4
Werkzeug==3.0.3
wsproto==1.2.0
